// src/types/vehicle-document.types.ts
export interface VehicleDocumentItem {
  id: number;
  registrationId: number;
  vehicleDataId?: number | null;
  dealerId?: number | null;
  regionId?: number | null;
  dealerName: string;
  stnkName: string;
  regionName: string;
  machineNumber: string;
  invoiceReceiveDate: string;
  bpkbRegistrationDate: string;
  stnkRegistrationDate: string;
  skpdPaymentDate: string;
  bpkbReceivedDate: string;
  stnkReceivedDate: string;
  skpdReceivedDate: string;
  tnkbReceivedDate: string;
  tnkbNumber: string;
  noticeFee: number;
  vendorEmployee: string;
  vehicleType: string;
}